backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
pillow==11.0.0
pytesseract==0.3.13
opencv-python-headless==4.10.0.84
pandas==2.2.3
rapidfuzz==3.11.0
openai==1.59.7
pypdf==5.1.0
PyMuPDF==1.24.14
beautifulsoup4==4.12.3
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
